Standardize file_handle type to int in save_keymap_to_file
Changed file_handle declaration from uint16_t to int in save_keymap_to_file
to match load_keymap_file and follow DOS INT 21h conventions.

File handles are signed 16-bit values where -1 (0xFFFF) can indicate error,
though this code checks the carry flag for error detection. Using int for
consistency across similar file I/O functions in the codebase.

+++ src/game_main.c
@@ -1071,7 +1071,7 @@ static const char* get_scancode_name(uint8_t scancode)
 static int save_keymap_to_file(uint8_t *keymap_data)
 {
     union REGS regs;
-    uint16_t file_handle;
+    int file_handle;
     uint16_t bytes_written;
     
     /* Create or truncate the file */
